move all errors and warnings to separate toplevel package
