+++ cassiopeia/mixins/warnings.py
@@ -1,0 +1,10 @@
+class CassiopeiaTreeWarning(UserWarning):
+    """A Warning for the CassiopeiaTree class."""
+
+    pass
+
+
+class SharedMutationJoiningSolverWarning(UserWarning):
+    """A warning class for SharedMutationJoiningSolver."""
+
+    pass
