+++ cassiopeia/simulator/TreeSimulator.py
@@ -7,14 +7,7 @@ implement a method called `simulate_tree`.
 import abc
 
 from cassiopeia.data import CassiopeiaTree
-
-
-class TreeSimulatorError(Exception):
-    """An Exception class for all exceptions generated by
-    TreeSimulator or a subclass of TreeSimulator
-    """
-
-    pass
+from cassiopeia.mixins import TreeSimulatorError
 
 
 class TreeSimulator(abc.ABC):
